fix: reduce runtime complexity of Delete() to O(1)
Use a doubly-linked list instead of a slice to store messages in the
queue. This reduces runtime complexity of Delete() from O(n) to O(1) at
the cost of increasing it from O(1) to O(n) on Messages().

The assumption is that Delete() is called more often and in more
performance critical code paths than Messages().